+++ project/Engine/Module/Render/RenderNode/Forward/Particle/ParticleBillboardNode/ParticleBillboardNode.cpp
@@ -1,32 +1,39 @@
-#include "LineNode.h"
+#include "ParticleBillboardNode.h"
 
 #include "Engine/Rendering/DirectX/PipelineState/PipelineState.h"
 #include "Engine/Rendering/DirectX/PipelineState/PSOBuilder/PSOBuilder.h"
 #include "Engine/Rendering/DirectX/DirectXResourceObject/DepthStencil/DepthStencil.h"
 
-LineNode::LineNode() = default;
+ParticleBillboardNode::ParticleBillboardNode() = default;
 
-LineNode::~LineNode() noexcept = default;
+ParticleBillboardNode::~ParticleBillboardNode() noexcept = default;
 
-void LineNode::initialize() {
+void ParticleBillboardNode::initialize() {
 	depthStencil = DepthStencilValue::depthStencil;
 	create_pipeline_state();
-	pipelineState->set_name("LineNode");
+	pipelineState->set_name("ParticleBillboardNode");
 	primitiveTopology = D3D10_PRIMITIVE_TOPOLOGY_TRIANGLELIST;
 }
 
-void LineNode::create_pipeline_state() {
+void ParticleBillboardNode::create_pipeline_state() {
 	RootSignatureBuilder rootSignatureBuilder;
+	rootSignatureBuilder.add_structured(D3D12_SHADER_VISIBILITY_VERTEX, 0, 1); // 0 :  transform
 	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_VERTEX, 0); // 1 : camera
-	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_PIXEL, 0); // 2 : material
+	rootSignatureBuilder.add_texture(D3D12_SHADER_VISIBILITY_PIXEL, 0); // 2 : texture
+	rootSignatureBuilder.sampler( // sampler
+		D3D12_SHADER_VISIBILITY_PIXEL,
+		0,
+		D3D12_FILTER_ANISOTROPIC
+	);
 
 	InputLayoutBuilder inputLayoutBuilder;
-	inputLayoutBuilder.add_element("POSITION", 0, DXGI_FORMAT_R32G32B32_FLOAT);
+	inputLayoutBuilder.add_element("POSITION", 0, DXGI_FORMAT_R32G32_FLOAT);
+	inputLayoutBuilder.add_element("TEXCOORD", 0, DXGI_FORMAT_R32G32_FLOAT);
 
-	ShaderBuilder shaderBuilder;
-	shaderBuilder.initialize(
-		"EngineResources/HLSL/Line/Line.VS.hlsl",
-		"EngineResources/HLSL/Line/Line.PS.hlsl"
+	ShaderBuilder shaderManager;
+	shaderManager.initialize(
+		"EngineResources/HLSL/Forward/Particle/ParticleBillboard/ParticleBillboard.VS.hlsl",
+		"EngineResources/HLSL/Forward/Particle/ParticleBillboard/ParticleBillboard.PS.hlsl"
 	);
 
 	std::unique_ptr<PSOBuilder> psoBuilder = std::make_unique<PSOBuilder>();
@@ -35,10 +42,11 @@ void LineNode::create_pipeline_state() {
 	psoBuilder->inputlayout(inputLayoutBuilder.build());
 	psoBuilder->rasterizerstate();
 	psoBuilder->rootsignature(rootSignatureBuilder.build());
-	psoBuilder->shaders(shaderBuilder);
+	psoBuilder->shaders(shaderManager);
 	psoBuilder->primitivetopologytype();
 	psoBuilder->rendertarget();
 
 	pipelineState = std::make_unique<PipelineState>();
 	pipelineState->initialize(psoBuilder->get_rootsignature(), psoBuilder->build());
+
 }
