+++ project/Engine/Module/Render/RenderNode/Forward/SkinningMesh/SkinningMeshNode.cpp
@@ -1,24 +1,24 @@
-#include "Object3DNode.h"
+#include "SkinningMeshNode.h"
 
 #include "Engine/Module/Render/RenderTargetGroup/SingleRenderTarget.h"
 #include "Engine/Rendering/DirectX/DirectXResourceObject/DepthStencil/DepthStencil.h"
 #include "Engine/Rendering/DirectX/DirectXResourceObject/OffscreenRender/OffscreenRender.h"
 #include "Engine/Rendering/DirectX/PipelineState/PipelineState.h"
 #include "Engine/Rendering/DirectX/PipelineState/PSOBuilder/PSOBuilder.h"
-
 #include "Engine/Rendering/RenderingSystemValues.h"
 
-Object3DNode::Object3DNode() = default;
-Object3DNode::~Object3DNode() noexcept = default;
+SkinningMeshNode::SkinningMeshNode() = default;
+SkinningMeshNode::~SkinningMeshNode() noexcept = default;
 
-void Object3DNode::initialize() {
+void SkinningMeshNode::initialize() {
 	depthStencil = DepthStencilValue::depthStencil;
 	create_pipeline_state();
-	pipelineState->set_name("Object3DNode");
+	pipelineState->set_name("SkinningMeshNode");
 	primitiveTopology = D3D10_PRIMITIVE_TOPOLOGY_TRIANGLELIST;
+
 }
 
-void Object3DNode::set_render_target(const std::shared_ptr<SingleRenderTarget>& renderTarget_) {
+void SkinningMeshNode::set_render_target(const std::shared_ptr<SingleRenderTarget>& renderTarget_) {
 	if (renderTarget_) {
 		renderTarget = renderTarget_;
 		resultSvtHandle = renderTarget_->offscreen_render().texture_gpu_handle();
@@ -33,13 +33,14 @@ void Object3DNode::set_render_target(const std::shared_ptr<SingleRenderTarget>& 
 	}
 }
 
-void Object3DNode::create_pipeline_state() {
+void SkinningMeshNode::create_pipeline_state() {
 	RootSignatureBuilder rootSignatureBuilder;
 	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_VERTEX, 0); // 0 :  transform
 	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_VERTEX, 1); // 1 : camera
 	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_PIXEL, 0); // 2 : material
 	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_PIXEL, 1); // 3 : light
-	rootSignatureBuilder.add_texture(D3D12_SHADER_VISIBILITY_PIXEL); // 4 : texture
+	rootSignatureBuilder.add_texture(D3D12_SHADER_VISIBILITY_PIXEL, 0); // 4 : texture
+	rootSignatureBuilder.add_structured(D3D12_SHADER_VISIBILITY_VERTEX, 1); // 5 : bone
 	rootSignatureBuilder.sampler( // sampler
 		D3D12_SHADER_VISIBILITY_PIXEL,
 		0,
@@ -47,14 +48,17 @@ void Object3DNode::create_pipeline_state() {
 	);
 
 	InputLayoutBuilder inputLayoutBuilder;
-	inputLayoutBuilder.add_element("POSITION", 0, DXGI_FORMAT_R32G32B32A32_FLOAT);
-	inputLayoutBuilder.add_element("TEXCOORD", 0, DXGI_FORMAT_R32G32_FLOAT);
-	inputLayoutBuilder.add_element("NORMAL", 0, DXGI_FORMAT_R32G32B32_FLOAT);
+	inputLayoutBuilder.add_element("POSITION", 0, DXGI_FORMAT_R32G32B32A32_FLOAT, 0);
+	inputLayoutBuilder.add_element("TEXCOORD", 0, DXGI_FORMAT_R32G32_FLOAT, 0);
+	inputLayoutBuilder.add_element("NORMAL", 0, DXGI_FORMAT_R32G32B32_FLOAT, 0);
+
+	inputLayoutBuilder.add_element("WEIGHT", 0, DXGI_FORMAT_R32G32B32A32_FLOAT, 1);
+	inputLayoutBuilder.add_element("INDEX", 0, DXGI_FORMAT_R32G32B32A32_UINT, 1);
 
 	ShaderBuilder shaderBuilder;
 	shaderBuilder.initialize(
-		"EngineResources/HLSL/3DObject/Object3d.VS.hlsl",
-		"EngineResources/HLSL/3DObject/Object3d.PS.hlsl"
+		"EngineResources/HLSL/Forward/AnimatedMesh/AnimatedMesh.VS.hlsl",
+		"EngineResources/HLSL/Forward/AnimatedMesh/AnimatedMesh.PS.hlsl"
 	);
 
 	std::unique_ptr<PSOBuilder> psoBuilder = std::make_unique<PSOBuilder>();
