+++ project/Engine/Module/Render/RenderNode/Posteffect/Grayscale/GrayscaleNode.cpp
@@ -1,4 +1,4 @@
-#include "ChromaticAberrationNode.h"
+#include "GrayscaleNode.h"
 
 #include "Engine/Rendering/DirectX/DirectXCommand/DirectXCommand.h"
 #include "Engine/Rendering/DirectX/PipelineState/PipelineState.h"
@@ -6,34 +6,33 @@
 
 #ifdef _DEBUG
 #include <imgui.h>
-#include "Engine/Application/EngineSettings.h"
 #endif // _DEBUG
 
-ChromaticAberrationNode::ChromaticAberrationNode() = default;
+GrayscaleNode::GrayscaleNode() = default;
 
-ChromaticAberrationNode::~ChromaticAberrationNode() noexcept = default;
+GrayscaleNode::~GrayscaleNode() = default;
 
-void ChromaticAberrationNode::initialize() {
+void GrayscaleNode::initialize() {
 	create_pipeline_state();
 	primitiveTopology = D3D10_PRIMITIVE_TOPOLOGY_TRIANGLELIST;
-	*aberrationLevel.get_data() = CVector2::ZERO;
+	*isGray.get_data() = false;
 }
 
-void ChromaticAberrationNode::draw() {
+void GrayscaleNode::draw() {
 	auto&& command = DirectXCommand::GetCommandList();
-	command->SetGraphicsRootConstantBufferView(0, aberrationLevel.get_resource()->GetGPUVirtualAddress());
 	command->SetGraphicsRootDescriptorTable(1, textureGPUHandle);
+	command->SetGraphicsRootConstantBufferView(0, isGray.get_resource()->GetGPUVirtualAddress());
 	command->DrawInstanced(3, 1, 0, 0);
 }
 
-void ChromaticAberrationNode::set_texture_resource(const D3D12_GPU_DESCRIPTOR_HANDLE& textureGPUHandle_) {
+void GrayscaleNode::set_texture_resource(const D3D12_GPU_DESCRIPTOR_HANDLE& textureGPUHandle_) {
 	textureGPUHandle = textureGPUHandle_;
 }
 
-void ChromaticAberrationNode::create_pipeline_state() {
+void GrayscaleNode::create_pipeline_state() {
 	RootSignatureBuilder rootSignatureBuilder;
 	rootSignatureBuilder.add_cbv(D3D12_SHADER_VISIBILITY_PIXEL, 0);
-	rootSignatureBuilder.add_texture(D3D12_SHADER_VISIBILITY_PIXEL, 0, 1);
+	rootSignatureBuilder.add_texture(D3D12_SHADER_VISIBILITY_PIXEL);
 	rootSignatureBuilder.sampler(
 		D3D12_SHADER_VISIBILITY_PIXEL,
 		0
@@ -42,7 +41,7 @@ void ChromaticAberrationNode::create_pipeline_state() {
 	ShaderBuilder shaderManager;
 	shaderManager.initialize(
 		"EngineResources/HLSL/FullscreenShader.hlsl",
-		"EngineResources/HLSL/ChromaticAberration/ChromaticAberration.PS.hlsl"
+		"EngineResources/HLSL/Posteffect/Grayscale/Grayscale.PS.hlsl"
 	);
 
 	std::unique_ptr<PSOBuilder> psoBuilder = std::make_unique<PSOBuilder>();
@@ -55,12 +54,10 @@ void ChromaticAberrationNode::create_pipeline_state() {
 
 	pipelineState = std::make_unique<PipelineState>();
 	pipelineState->initialize(psoBuilder->get_rootsignature(), psoBuilder->build());
-
 }
 
 #ifdef _DEBUG
-void ChromaticAberrationNode::debug_gui() {
-	ImGui::DragFloat("AberrationLevelX", &aberrationLevel.get_data()->x, 0.1f / EngineSettings::CLIENT_WIDTH, -0.5f, 0.5f, "%.4f");
-	ImGui::DragFloat("AberrationLevelY", &aberrationLevel.get_data()->y, 0.1f / EngineSettings::CLIENT_HEIGHT, -0.5f, 0.5f, "%.4f");
+void GrayscaleNode::debug_gui() {
+	ImGui::Checkbox("IsGray", reinterpret_cast<bool*>(isGray.get_data()));
 }
 #endif // _DEBUG
