+++ project/Engine/Module/Render/RenderNode/Posteffect/Outline/OutlineNode.cpp
@@ -46,7 +46,7 @@ void OutlineNode::create_pipeline_state() {
 	ShaderBuilder shaderManager;
 	shaderManager.initialize(
 		"EngineResources/HLSL/FullscreenShader.hlsl",
-		"EngineResources/HLSL/Outline/Outline.PS.hlsl"
+		"EngineResources/HLSL/Posteffect/Outline/Outline.PS.hlsl"
 	);
 
 	std::unique_ptr<PSOBuilder> psoBuilder = std::make_unique<PSOBuilder>();
