+++ project/EngineResources/HLSL/Debug/NormalColor/NormalColor.PS.hlsl
