+++ project/EngineResources/HLSL/Debug/NormalColor/NormalColor.VS.hlsl
