+++ project/EngineResources/HLSL/Forward/3DObject/Object3d.VS.hlsl
