+++ project/EngineResources/HLSL/Forward/AnimatedMesh/AnimatedMesh.VS.hlsl
