+++ project/EngineResources/HLSL/Forward/AnimatedMesh/AnimatedMesh.hlsli
