+++ project/EngineResources/HLSL/Posteffect/Grayscale/Grayscale.hlsli
